refactor : view 클래스 패키징
