+++ src/main/java/racinggame/view/InputView.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -8,7 +8,7 @@ import java.util.List;
 import racinggame.domain.Car;
 import racinggame.domain.Round;
 
-class InputView {
+public class InputView {
 
     public List<Car> readCars() throws IOException {
         String inputNames = readLine("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
